requirements.txt
aiohttp
pyrotgfork[fast]
python-dotenv
